Move to C++17 and std::filesystem.
With this, we are able to get rid of all the increasingly broken file
system utility code, and trust std::filesystem to handle all the cross-
platform complexity.

Unfortunately std::filesystem support remains a little elusive; it is
well supported in Visual Studio (especially 2019), but not by default
in Mac OS X, and even in GCC 8.0 it requires an explicit '-l c++fs'.

This also silences some of the more egregious compiler warnings, mostly
by being explicit about where we cast to the uint32 types glTF prefers.

+++ src/raw/RawModel.cpp
@@ -632,7 +632,7 @@ void RawModel::CreateMaterialModels(
   }
 }
 
-int RawModel::GetNodeById(const long nodeId) const {
+int RawModel::GetNodeById(const uint32_t nodeId) const {
   for (size_t i = 0; i < nodes.size(); i++) {
     if (nodes[i].id == nodeId) {
       return (int)i;
@@ -641,7 +641,7 @@ int RawModel::GetNodeById(const long nodeId) const {
   return -1;
 }
 
-int RawModel::GetSurfaceById(const long surfaceId) const {
+int RawModel::GetSurfaceById(const uint32_t surfaceId) const {
   for (size_t i = 0; i < surfaces.size(); i++) {
     if (surfaces[i].id == surfaceId) {
       return (int)i;
